Workflow: Support pushing images in PR's, support canary tags
With this change it's no longer necessary to bring test code onto the
main branch and to create releases of images that not actually are
releases.

In PR's the workflow now also publishes images tagged with the PR
version number, e.g. user/image:pr123.

Also, every change that gets merged to the main branch will be tagged
with the 'canary' tag. The 'latest' tags are only used for releases.