Provide custom scripts for watching configs, install inotify-tools
In order to be able to automatically reload the Apache webserver
gracefully when Apache config changes, a custom script for watching the
config has been provided. This script is now started as a background
process by a modified run script. Like this, the config watcher
automatically gets started everytime the Apache webserver gets started
too.

As the config watcher is based on the inotifywait command, the package
inotify-tools has been installed in the container.